requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1.3
django-chartjs==2.3.0
django-heroku==0.3.1
psycopg2==2.9.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
validate-email==1.3
waitress==2.1.2
whitenoise==6.2.0
xlwt==1.3.0
